+++ lib/dartapi.dart
@@ -2,4 +2,6 @@ export 'cli/create_project.dart';
 export 'cli/run_server.dart';
 export 'cli/generate_controllers.dart';
 export 'cli/generate_docs.dart';
+export 'cli/generate_migration.dart';
+export 'cli/run_migrations.dart';
 export 'constants/create_command_constants.dart';